requirements.txt
absl-py==1.3.0
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
boto3==1.24.85
botocore==1.27.85
bravado==11.0.3
bravado-core==5.17.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.7
executing==1.2.0
flatbuffers==22.9.24
fonttools==4.37.1
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
ipython==8.8.0
jedi==0.18.2
jmespath==1.0.1
joblib==1.1.0
jsonpointer==2.3
jsonref==0.2
jsonschema==3.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
monotonic==1.6
msgpack==1.0.4
numpy==1.23.4
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pycparser==2.21
Pygments==2.14.0
PyJWT==2.5.0
pyotp==2.8.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3987==1.3.8
robin-stocks==2.1.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.6.2
strict-rfc3339==0.7
swagger-spec-validator==2.7.6
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-model-optimization==0.7.3
termcolor==2.0.1
threadpoolctl==3.1.0
traitlets==5.8.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webcolors==1.12
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0

tensorflow~=2.11.0
nltk~=3.8.1
finvizfinance~=0.14.5
DateTime~=5.0
neptune~=1.1.1